backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.20
langgraph==0.0.20
langsmith==0.0.83

# Upstage Solar API
langchain-upstage==0.1.0

# Vector Store
faiss-cpu==1.7.4
numpy==1.26.3

# Database (mock for MVP)
# psycopg2-binary==2.9.9
# redis==5.0.1

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
